replaced all big decimals with ints

+++ src/main/java/checkoutkata/PriceProvider.java
@@ -1,9 +1,7 @@
 package checkoutkata;
 
-import java.math.BigDecimal;
-
 public interface PriceProvider {
 
-    public BigDecimal getPrice(char sku);
+    public int getPrice(char sku);
 
 }
